+++ src/main/java/com/tsong/cmall/service/CarouselService.java
@@ -16,7 +16,7 @@ public interface CarouselService {
 
     Carousel getCarouselById(Integer id);
 
-    Boolean deleteBatch(Integer[] ids);
+    Boolean deleteBatch(Long[] ids);
 
     /**
      * @Description 返回固定数量的轮播图对象(首页调用)
